==> 01_mortality_model/1_DataCleaning.Rmd <==
---
title: '**TFM: Data Cleaning**'
author: '*Roberto Jesús Alcaraz Molina*'
date: "11/02/2021"
output:
  prettydoc::html_pretty:
    theme: cayman
    df_print: paged
header-includes: 
- \usepackage{float}
- \usepackage{amsbsy}
- \usepackage{amsmath}
- \usepackage{graphicx}
- \usepackage{subfig}
- \usepackage{booktabs}
---

```{r include=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      echo = T, 
                      message = FALSE,
                      fig.pos="H", 
                      fig.align="center",
                      fig.width=15,
                      cache=FALSE, error = TRUE)
```

```{r, echo = F}
# install.packages("pacman")
pacman::p_load(tidyverse, tidymodels)
```

```{r}
# insurance <- read.table("BBDD_DEFINITIVA_V4.txt", header = T)
# saveRDS(insurance, "BBDD_DEFINITIVA_V4.RDS")

insurance <- readRDS("../00_data/BBDD_DEFINITIVA_V4.RDS")
skimr::skim(insurance)
```

- CLAVE: Identification of the person
- PERIODO: Period (year)
- COVER: Number of coverages (1, 2 or 3)
- SEXO: 
- FUMADOR:       
- BUENA_SALUD: Yes or No  
- EDAD_ACTUARIAL: actuarial age in each year
- EDAD_SUS: actuarial age in the moment of suscription  
- EXP: Tiempo de exposición de la póliza al riesgo de fallecimiento durante el año.
- mortality: Yes or No

- DURACION: Number of years in force of the policy
- DURACION_CAT: Rounded duration

- CAPITAL: Capital asegurado. Dinero que recibirá el beneficiario en caso de mortality

- CAPITAL_CAT1: 
- CAPITAL_CAT2:      
- CAPITAL_CAT3: 
- CAPITAL_CAT1_DESC: 
- CAPITAL_CAT2_DESC:
- CAPITAL_CAT3_DESC:

- BMI:          
- BMI_CAT1:    
- BMI_CAT2:          
- BMI_CAT3:
- BMI_CAT1_DESC:    
- BMI_CAT2_DESC:  

```{r}
insurance <- insurance %>%
  rename(
    key              = CLAVE, 
    actuarial_age    = EDAD_ACTUARIAL, 
    subs_age          = EDAD_SUS,
    duration_cat     = DURACION_CAT,
    capital          = CAPITAL,
    ) %>%
  mutate(
    period           = factor(insurance$PERIODO),
    cover            = factor(insurance$COVER),
    sex              = factor(ifelse(insurance$SEXO == 0, "man", "woman")),
    smoker           = factor(ifelse(insurance$FUMADOR == 0, "no", "yes")),
    good_health      = factor(ifelse(insurance$BUENA_SALUD == 1, "yes", "no")),
    exp              = as.numeric(gsub(",", ".", gsub("\\.", "", insurance$EXP))),
    mortality        = factor(ifelse(insurance$SINIESTRO == 1, "yes", "no")),
    duration         = as.numeric(gsub(",",".", gsub("\\.", "", insurance$DURACION))),
    capital_factor_1 = factor(insurance$CAPITAL_CAT1_DESC, 
                              levels = c("0-30.000", "30.001-60.000", 
                                         "60.001-90.000","90.001-120.000", 
                                         "120.001-150.000", "+150.000"),
                              ordered = T),
    capital_factor_2 = factor(insurance$CAPITAL_CAT2_DESC, 
                              levels = c("0-60.000", "60.001-120.000", 
                                         "+120.000"),
                              ordered = T),
    capital_factor_3 = factor(insurance$CAPITAL_CAT3_DESC, 
                              levels = c("0-100.000", "100.001-200.000", 
                                         "+200.000"),
                              ordered = T),
    BMI              = as.numeric(gsub(",", ".", gsub("\\.", "", insurance$IMC))),
    BMI_factor_1     = factor(insurance$IMC_CAT1_DESC, 
                              levels = c("Normal", "Riesgo"),
                              labels = c("normal", "risk")),
    BMI_factor_2     = factor(insurance$IMC_CAT2_DESC,
                              levels = c("Normal", "Sobrepeso"),
                              labels = c("normal", "overweight"))
  ) %>%
  
  select(
    key, period, cover, sex,  smoker, good_health, actuarial_age, subs_age, exp,
    mortality, duration, duration_cat, capital, capital_factor_1, capital_factor_2,
    capital_factor_3, BMI, BMI_factor_1, BMI_factor_2
  )
```

Here, we remove the repeated columns
```{r}
insurance <- tibble::rowid_to_column(insurance, "ID")

# here we take the IDs that are not repeated
IDs <- insurance %>%
  arrange(desc(mortality)) %>%
  distinct_at(vars(key, period, cover), .keep_all = T) %>%
  select(ID)

# here we ensure that the rows we are removing have mortality = 0
insurance %>%
  filter(!(ID %in% IDs$ID)) %>%
  select(mortality) %>%
  summary()

insurance <- insurance %>%
  filter(ID %in% IDs$ID)

insurance
```

Also, we create another df of only the people that had some sinister
```{r}
# mortalityS
key_mortality <- insurance %>% # tomamos todas las claves de las personas con mort
  filter(mortality == "yes") %>%
  select(key)

insurance_mortality <- insurance %>%
  filter(key %in% key_mortality$key)

insurance_mortality
```

And another one with the people that drop
```{r}
# lapse
# Here we take the keys of the lapse

keys_lapse <- insurance %>%
  filter((key %in% key_mortality$key) == F) %>%
  group_by(key) %>%
  count() %>%
  mutate(lapse = ifelse(n == 1 | n == 2, "yes", "no")) %>%
  filter(lapse == "yes") %>%
  select(key)

insurance_lapse <- insurance %>%
  filter((key %in% key_mortality$key) == F) %>%
  mutate(lapse = ifelse(key %in% keys_lapse$key, "yes", "no")) %>%
  arrange(key) 

for (i in 2:nrow(insurance_lapse)){
  if (insurance_lapse$lapse[i-1] == "yes") {
    if (insurance_lapse$key[i-1] == insurance_lapse$key[i]) {
      insurance_lapse$lapse[i-1] = "no"
    }
  }
}

insurance_lapse$mortality = NULL
insurance_lapse
```

```{r}
saveRDS(insurance, file = "../00_data/insurance.RDS")
saveRDS(insurance_lapse, file = "../00_data/insurance_lapse.RDS")

insurance <- readRDS("../00_data/insurance.RDS")
insurance_lapse <- readRDS("../00_data/insurance_lapse.RDS")
```
