==> 01_mortality_model/2_EDA_mortality.Rmd <==
---
title: '**Mortality: Exploratory Data Analysis (EDA)**'
author: '*Roberto Jesús Alcaraz Molina*'
date: "14/04/2021"
output:
  prettydoc::html_pretty:
    theme: cayman
    df_print: paged
header-includes: 
- \usepackage{float}
- \usepackage{amsbsy}
- \usepackage{amsmath}
- \usepackage{graphicx}
- \usepackage{subfig}
- \usepackage{booktabs}
---

```{r include=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      echo = T, 
                      message = FALSE,
                      fig.pos="H", 
                      fig.align="center",
                      fig.width=15,
                      cache=FALSE, error = TRUE)
```

```{r, echo = F}
# install.packages("pacman")
pacman::p_load(tidyverse, tidymodels, workflows, tune, patchwork, FactoMineR, dotwhisker, splines, ggpubr)
theme_set(theme_bw())

# Models packages
pacman::p_load(ranger)
```

```{r}
insurance <- readRDS("../00_data/insurance.RDS")
```


# Description and visualization of the variables

Our variables are:

- **Indicative variables**:

  - `ID`
  - `key`

- **Categorical variables (10 + 1)**:

  - `mortality` (target)
  - `period`
  - `cover`
  - `sex`
  - `smoker`
  - `good_health`
  - `capital_factor_1`
  - `capital_factor_2`
  - `capital_factor_3`
  - `BMI_factor_1`
  - `BMI_factor_2`


- **Numerical variables (7)**:

  - `actuarial_age`
  - `sus_age`
  - `exp`
  - `duration`
  - `duration_cat`
  - `capital`
  - `BMI`

## Target variable: `mortality`

```{r}
# mortality
table(insurance$mortality)/nrow(insurance)
```

We are going to split the data in 70% for training and 30% for testing
```{r}
set.seed(123)

insurance_split <- initial_split(insurance, prop = 0.7, strata = mortality)
insurance_train <- training(insurance_split)
insurance_test  <- testing(insurance_split)
```

## Categorical variables (10 + 1)

- `period`
- `cover`
- `sex`
- `smoker`
- `good_health`
- `capital_factor_1`
- `capital_factor_2`
- `capital_factor_3`
- `BMI_factor_1`
- `BMI_factor_2`



### Cover
```{r}
table(insurance_train$cover)/nrow(insurance_train)

insurance_train %>%
  ggplot(aes(x = cover)) +
  geom_bar(aes(y=..count../sum(..count..)),color = "black", fill = "indianred", alpha = 0.8) +
  ylab("") +
  scale_y_continuous(labels = percent_format())

tab <- table(insurance_train$cover, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

# as we increase the number of coverages, the proportion increases
```

### sex
```{r}
table(insurance_train$sex)/nrow(insurance_train)

insurance_train %>%
  ggplot(aes(x = sex)) +
  geom_bar(aes(y=..count../sum(..count..)),color = "black", fill = "indianred", alpha = 0.8) +
  ylab("") +
  scale_y_continuous(labels = percent_format())

tab <- table(insurance_train$sex, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

# more than double in men
```

### smoker
```{r}
table(insurance_train$smoker)/nrow(insurance_train)

insurance_train %>%
  ggplot(aes(x = smoker)) +
  geom_bar(aes(y=..count../sum(..count..)),color = "black", fill = "indianred", alpha = 0.8) +
  ylab("") +
  scale_y_continuous(labels = percent_format())

tab <- table(insurance_train$smoker, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

# double in smokers
```

### good_health 
```{r}
table(insurance_train$good_health)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = good_health)) +
  geom_bar(aes(y=..count../sum(..count..)),color = "black", fill = "indianred", alpha = 0.8) +
  ylab("") +
  scale_y_continuous(labels = percent_format())
tab <- table(insurance_train$good_health, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

# slightly higher for the people with good health
```

### capital_DESC
```{r}
table(insurance_train$capital_factor_1)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = capital_factor_1)) +
  geom_bar() 

tab <- table(insurance_train$capital_factor_1, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

table(insurance_train$capital_factor_2)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = capital_factor_2)) +
  geom_bar() 
tab <- table(insurance_train$capital_factor_2, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))
# we should convert it to dummy. Lower proportion of mortality in the last group

table(insurance_train$capital_factor_3)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = capital_factor_3)) +
  geom_bar() 
tab <- table(insurance_train$capital_factor_3, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))
# convert it to dummy
```

### BMI_DESC

```{r}
table(insurance_train$BMI_factor_1)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = BMI_factor_1)) +
  geom_bar() 

tab <- table(insurance_train$BMI_factor_1, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

table(insurance_train$BMI_factor_2)/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = BMI_factor_2)) +
  geom_bar() 

tab <- table(insurance_train$BMI_factor_2, insurance_train$mortality)
cbind(tab, tab[, 2] / (tab[, 1] + tab[, 2]))

# the two of them are very similar. Double in people with risk/overweight
```

## Numerical variables (7)

- `actuarial_age`
- `sus_age`
- `exp`
- `duration`
- `duration_cat`
- `capital`
- `BMI`

### actuarial and suscription age
```{r}
insurance_train %>%
  ggplot(aes(x = actuarial_age)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white", bins = 20)+
  geom_density(alpha=.5, fill="indianred") +
  scale_y_continuous(labels = percent_format()) +
  ylab("")
# seems good

t1 <- insurance_train %>%
  group_by(actuarial_age) %>%
  mutate(mortality = sum(mortality == "yes")) %>%
  select(actuarial_age, mortality) %>%
  distinct() %>%
  arrange(actuarial_age)

t2 <- insurance_train %>%
  group_by(actuarial_age) %>%
  mutate(exp = sum(exp)) %>%
  select(actuarial_age, exp) %>%
  distinct() %>%
  arrange(actuarial_age)

t <- t1 %>%
  bind_cols(exp = t2$exp, .name_repair = "unique") %>%
  mutate(`mortality/exp` = mortality/exp)

t %>%
  ggplot(aes(actuarial_age, `mortality/exp`)) +
  geom_point() +
  geom_smooth(method = "loess", color = "indianred", fill = "indianred") +
  scale_y_continuous(labels = percent_format())

quantile(insurance_train$actuarial_age, probs = c(0.01, 0.99))
```

```{r}
insurance_train %>%
  ggplot(aes(x = subs_age)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white")+
  geom_density(alpha=.5, fill="indianred") +
  scale_y_continuous(labels = percent_format()) +
  ylab("")
```



### exp

```{r}
table(ifelse(insurance_train$exp == 1, "=1", "<1"))/nrow(insurance_train)
insurance_train %>%
  ggplot(aes(x = exp)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white")+
  geom_density(alpha=.2, fill="#FF6666")
```

### duration
```{r}
insurance_train %>%
  ggplot(aes(x = duration)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white", bins = 20)+
  geom_density(alpha=.5, fill="indianred") +
  scale_y_continuous(labels = percent_format()) +
  ylab("")

t1 <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(mortality = sum(mortality == "yes")) %>%
  select(duration, mortality) %>%
  distinct() %>%
  arrange(duration)

t2 <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(exp = sum(exp)) %>%
  select(duration, exp) %>%
  distinct() %>%
  arrange(duration)

t <- t1 %>%
  bind_cols(exp = t2$exp, .name_repair = "unique") %>%
  mutate(`mortality/exp` = mortality/exp)

t %>%
  ggplot(aes(duration, `mortality/exp`)) +
  geom_point() +
  geom_smooth(method = "loess", color = "indianred", fill = "indianred") +
  scale_y_continuous(labels = percent_format())
```
### Capital
```{r}
summary(insurance_train$capital)
p1 <- insurance_train %>%
  ggplot(aes(x = capital)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white", bins = 20)+
  geom_density(alpha=.5, fill="indianred") +
  scale_y_continuous(labels = percent_format()) +
  ylab("")

p2 <- insurance_train %>%
  ggplot(aes(x = capital)) +
  geom_boxplot()

ggarrange(p1, p2, ncol = 1)

insurance_train %>%
  filter(capital > 200000)

insurance_train %>%
  ggplot(aes(x = capital)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white")+
  geom_density(alpha=.2, fill="#FF6666") +
  scale_x_log10()
# here we have some outliers. We should take it into account.
```


### BMI
```{r}
summary(insurance_train$BMI)

insurance_train %>%
  ggplot(aes(x = BMI)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white", bins = 20)+
  geom_density(alpha=.5, fill="indianred") +
  scale_y_continuous(labels = percent_format()) +
  ylab("") 

insurance_train %>%
  ggplot(aes(x = BMI)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white")+
  geom_density(alpha=.2, fill="#FF6666") +
  scale_x_log10()

# Some outliers too

t1 <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(mortality = sum(mortality == "yes")) %>%
  select(duration, mortality) %>%
  distinct() %>%
  arrange(duration)

t2 <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(exp = sum(exp)) %>%
  select(duration, exp) %>%
  distinct() %>%
  arrange(duration)

t <- t1 %>%
  bind_cols(exp = t2$exp, .name_repair = "unique") %>%
  mutate(`mortality/exp` = mortality/exp)

t %>%
  ggplot() +
  geom_point(aes(duration, `mortality/exp`), alpha = 0.4)
```

```{r}
names(insurance_train)

(a <- cor(insurance_train[, c(8, 12, 14, 18)]))
GGally::ggcorr(a, label = T,label_round = 2, name = "Correlation", label_alpha = T, layout.exp = 1)

insurance_train %>%
  ggplot(aes(x = capital, y = BMI)) +
  geom_point()
# we have no linear correlation between the numerical variables
```
