==> 01_mortality_model/3_FeatureEngineering_mortality.Rmd <==
---
title: '**Mortality: Feature Engineering**'
author: '*Roberto Jesús Alcaraz Molina*'
date: "14/04/2021"
output:
  prettydoc::html_pretty:
    theme: cayman
    df_print: paged
header-includes: 
- \usepackage{float}
- \usepackage{amsbsy}
- \usepackage{amsmath}
- \usepackage{graphicx}
- \usepackage{subfig}
- \usepackage{booktabs}
---

```{r include=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      echo = T, 
                      message = FALSE,
                      fig.pos="H", 
                      fig.align="center",
                      fig.width=10,
                      cache=FALSE, error = TRUE)
```

```{r, echo = F}
# install.packages("pacman")
# devtools::install_github("stevenpawley/recipeselectors")
pacman::p_load(tidyverse, tidymodels, workflowsets, tune, patchwork, dotwhisker, doParallel, mgcv, performance, recipeselectors, vip, ggfortify, ggpubr, cowplot, itsadug, mgcViz)
theme_set(theme_bw())

pacman::p_load(ranger)

scale_factor()
```

# Code
```{r}
insurance <- readRDS("../00_data/insurance.RDS")

insurance <- insurance %>%
  filter(actuarial_age >= 21, actuarial_age <= 61) %>%
  filter(duration <= 13)


set.seed(123)

insurance_split <- initial_split(insurance, prop = 0.7, strata = mortality)
insurance_train <- training(insurance_split)
insurance_test  <- testing(insurance_split)
```


In this step of the project, we will analyze the following things:

  - Variable selection through gam, drop (p-value > 0.2) and random forest.
  - Interaction between variables.
  - Preprocessing steps taking into account the EDA.

# Variable selection

## Recursive feature selection with Random Forest


```{r}
prima <- read.csv("../00_data/prima.csv", header = T, sep = ";")
colnames(prima) <- c("actuarial_age", "qx_men", "qy_women")
prima$qx_men <- as.numeric(gsub(",", ".", prima$qx_men))
prima$qy_women <- as.numeric(gsub(",", ".", prima$qy_women))
prima <- prima %>% 
  filter(actuarial_age > 20, actuarial_age < 62)
```

```{r}
insurance <- merge(insurance, prima, by = "actuarial_age") %>% 
  mutate(qx = if_else(sex == "man", qx_men, qy_women)) %>% 
  select(-qx_men, -qy_women)

insurance %>% 
  mutate(prima = qx * capital)
```



```{r}
rfe_model <- rand_forest() %>%
  set_engine("ranger", importance = "permutation") %>%
  set_mode("classification")

# Predictive model for the deaths
set.seed(1234)
select_rec <- 
  recipe(mortality ~ cover + sex + smoker + good_health +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  themis::step_upsample(mortality, over_ratio = 0.01) %>%
  step_log(capital, BMI, base = 10) %>%
  step_select_vip(all_predictors(), outcome = "mortality", model = rfe_model, top_p = 6)

new_data <- select_rec %>% prep() %>% juice()
new_data

prep_data <- 
  recipe(mortality ~ cover + sex + smoker + good_health +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  themis::step_upsample(mortality, over_ratio = 0.01) %>%
  step_log(capital, BMI, base = 10) %>% prep() %>% juice()

vip_model <- rfe_model %>% 
  fit(mortality ~ cover + sex + smoker + good_health +
                   actuarial_age + duration + capital + BMI,
         data = prep_data)

vip(vip_model, aesthetics = list(color = "black", fill = "indianred", alpha = 0.8))
# The recursive feature selection with Random Forest discard the variables cover and good_health.
```



## GAM models

```{r}
insurance_train_prep <- 
  recipe(mortality ~ sex + smoker + exp +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  themis::step_upsample(mortality, over_ratio = 0.01, seed = 123) %>%
  prep() %>% juice()
```


```{r, eval = F}
# Here we test if they are non-linear or not
gam_model <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps") + s(duration, bs = "ps") +
                   s(capital, bs = "ps") + s(BMI, bs = "ps"),
                 data = insurance_train_prep, family = "binomial", method = "REML", select = F)
summary(gam_model)

# Parametric coefficients:
#             Estimate Std. Error  z value Pr(>|z|)    
# (Intercept) -4.90664    0.03579 -137.079   <2e-16 ***
# sexwoman    -0.87869    0.05866  -14.978   <2e-16 ***
# smokeryes    0.40954    0.04573    8.955   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Approximate significance of smooth terms:
#                    edf Ref.df Chi.sq p-value    
# s(actuarial_age) 3.629  4.358 2008.4  <2e-16 ***
# s(duration)      6.851  7.473  154.8  <2e-16 ***
# s(capital)       7.520  7.885  139.6  <2e-16 ***
# s(BMI)           6.413  6.791  153.6  <2e-16 ***

# And here, if they should be in the model
gam_model2 <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps") + s(duration, bs = "ps") +
                   s(capital, bs = "ps") + s(BMI, bs = "ps"),
                 data = insurance_train_prep, family = "binomial", method = "REML", select = T)
summary(gam_model2)


# Parametric coefficients:
#             Estimate Std. Error  z value Pr(>|z|)    
# (Intercept) -4.87503    0.03456 -141.054   <2e-16 ***
# sexwoman    -0.89695    0.05831  -15.382   <2e-16 ***
# smokeryes    0.40570    0.04572    8.874   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Approximate significance of smooth terms:
#                    edf Ref.df  Chi.sq p-value    
# s(actuarial_age) 3.640      9 2010.85  <2e-16 ***
# s(duration)      7.471      9  167.82  <2e-16 ***
# s(capital)       7.096      9  143.33  <2e-16 ***
# s(BMI)           4.633      9   93.78  <2e-16 ***

plot(gam_model2, ylim = c(-5, 5))
```




# With interactions:

```{r}
# actuarial age with sex
t1 <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  mutate(
    mortality_man = sum(mortality == "yes" & sex == "man"),
    mortality_woman = sum(mortality == "yes" & sex == "woman"),
    ) %>%
  select(actuarial_age, mortality_man, mortality_woman) %>%
  distinct() %>%
  arrange(actuarial_age)

t_man <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  filter(sex == "man") %>%
  mutate(exp_man = sum(exp)) %>%
  select(actuarial_age, exp_man) %>%
  distinct() %>%
  arrange(actuarial_age)

t_woman <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  filter(sex == "woman") %>%
  mutate(exp_woman = sum(exp)) %>%
  select(actuarial_age, exp_woman) %>%
  distinct() %>%
  arrange(actuarial_age)

t <- t1 %>%
  bind_cols(
    exp_man = t_man$exp_man,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_woman = t_woman$exp_woman,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_man/exp` = mortality_man/exp_man,
    `mortality_woman/exp` = mortality_woman/exp_woman
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_man/exp`, `mortality_woman/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(actuarial_age, value, color = values, fill = values), alpha = 0.4)

group_colors <- c("indianred", "cornflowerblue")
t %>%
  ggplot(aes(actuarial_age, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# actuarial_age with smoker
t1 <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  mutate(
    mortality_smoker = sum(mortality == "yes" & smoker == "yes"),
    mortality_nonsmoker = sum(mortality == "yes" & smoker == "no"),
    ) %>%
  select(actuarial_age, mortality_smoker, mortality_nonsmoker) %>%
  distinct() %>%
  arrange(actuarial_age)

t_yes <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  filter(smoker == "yes") %>%
  mutate(exp_yes = sum(exp)) %>%
  select(actuarial_age, exp_yes) %>%
  distinct() %>%
  arrange(actuarial_age)

t_no <- insurance_train_prep %>%
  group_by(actuarial_age) %>%
  filter(smoker == "no") %>%
  mutate(exp_no = sum(exp)) %>%
  select(actuarial_age, exp_no) %>%
  distinct() %>%
  arrange(actuarial_age)

t <- t1 %>%
  bind_cols(
    exp_yes = t_yes$exp_yes,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_no = t_no$exp_no,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_smoker/exp` = mortality_smoker/exp_yes,
    `mortality_nonsmoker/exp` = mortality_nonsmoker/exp_no
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_smoker/exp`, `mortality_nonsmoker/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(actuarial_age, value, color = values, fill = values), alpha = 0.4)

t %>%
  ggplot(aes(actuarial_age, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# duration with sex
t1 <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(
    mortality_man = sum(mortality == "yes" & sex == "man"),
    mortality_woman = sum(mortality == "yes" & sex == "woman"),
    ) %>%
  select(duration, mortality_man, mortality_woman) %>%
  distinct() %>%
  arrange(duration)

t_man <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  filter(sex == "man") %>%
  mutate(exp_man = sum(exp)) %>%
  select(duration, exp_man) %>%
  distinct() %>%
  arrange(duration)

t_woman <- insurance_train %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  filter(sex == "woman") %>%
  mutate(exp_woman = sum(exp)) %>%
  select(duration, exp_woman) %>%
  distinct() %>%
  arrange(duration)

t <- t1 %>%
  bind_cols(
    exp_man = t_man$exp_man,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_woman = t_woman$exp_woman,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_man/exp` = mortality_man/exp_man,
    `mortality_woman/exp` = mortality_woman/exp_woman
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_man/exp`, `mortality_woman/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(duration, value, color = values, fill = values), alpha = 0.4)

t %>%
  ggplot(aes(duration, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# duration with smoker
t1 <- insurance_train_prep %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  mutate(
    mortality_smoker = sum(mortality == "yes" & smoker == "yes"),
    mortality_nonsmoker = sum(mortality == "yes" & smoker == "no"),
    ) %>%
  select(duration, mortality_smoker, mortality_nonsmoker) %>%
  distinct() %>%
  arrange(duration)

t_yes <- insurance_train_prep %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  filter(smoker == "yes") %>%
  mutate(exp_yes = sum(exp)) %>%
  select(duration, exp_yes) %>%
  distinct() %>%
  arrange(duration)

t_no <- insurance_train_prep %>%
  mutate(duration = round(duration)) %>%
  group_by(duration) %>%
  filter(smoker == "no") %>%
  mutate(exp_no = sum(exp)) %>%
  select(duration, exp_no) %>%
  distinct() %>%
  arrange(duration)

t <- t1 %>%
  bind_cols(
    exp_yes = t_yes$exp_yes,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_no = t_no$exp_no,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_smoker/exp` = mortality_smoker/exp_yes,
    `mortality_nonsmoker/exp` = mortality_nonsmoker/exp_no
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_smoker/exp`, `mortality_nonsmoker/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(duration, value, color = values, fill = values), alpha = 0.4)

t %>%
  ggplot(aes(duration, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# capital with sex
t1 <- insurance_train %>%
  mutate(capital = round(capital)) %>%
  group_by(capital) %>%
  mutate(
    mortality_man = sum(mortality == "yes" & sex == "man"),
    mortality_woman = sum(mortality == "yes" & sex == "woman"),
    ) %>%
  select(capital, mortality_man, mortality_woman) %>%
  distinct() %>%
  arrange(capital)

t1 <- t1[-5, ]

t_man <- insurance_train %>%
  mutate(capital = round(capital)) %>%
  group_by(capital) %>%
  filter(sex == "man") %>%
  mutate(exp_man = sum(exp)) %>%
  select(capital, exp_man) %>%
  distinct() %>%
  arrange(capital)

t_man <- t_man[-5, ]

t_woman <- insurance_train %>%
  mutate(capital = round(capital)) %>%
  group_by(capital) %>%
  filter(sex == "woman") %>%
  mutate(exp_woman = sum(exp)) %>%
  select(capital, exp_woman) %>%
  distinct() %>%
  arrange(capital)

t <- t1 %>%
  bind_cols(
    exp_man = t_man$exp_man,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_woman = t_woman$exp_woman,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_man/exp` = mortality_man/exp_man,
    `mortality_woman/exp` = mortality_woman/exp_woman
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_man/exp`, `mortality_woman/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(capital, value, color = values, fill = values), alpha = 0.4)

t %>%
  ggplot(aes(capital, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# capital with smoker
t1 <- insurance_train %>%
  mutate(capital = quantile(capital, probs = seq(0.01, 0.99, 0.01))) %>%
  group_by(capital) %>%
  mutate(
    mortality_smoker = sum(mortality == "yes" & smoker == "yes"),
    mortality_nonsmoker = sum(mortality == "yes" & smoker == "no"),
    ) %>%
  select(capital, mortality_smoker, mortality_nonsmoker) %>%
  distinct() %>%
  arrange(capital)

t_yes <- insurance_train %>%
  mutate(capital = round(capital)) %>%
  group_by(capital) %>%
  filter(smoker == "yes") %>%
  mutate(exp_yes = sum(exp)) %>%
  select(capital, exp_yes) %>%
  distinct() %>%
  arrange(capital)

t_no <- insurance_train %>%
  mutate(capital = round(capital)) %>%
  group_by(capital) %>%
  filter(smoker == "no") %>%
  mutate(exp_no = sum(exp)) %>%
  select(capital, exp_no) %>%
  distinct() %>%
  arrange(capital)

t <- t1 %>%
  bind_cols(
    exp_yes = t_yes$exp_yes,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_no = t_no$exp_no,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_smoker/exp` = mortality_smoker/exp_yes,
    `mortality_nonsmoker/exp` = mortality_nonsmoker/exp_no
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_smoker/exp`, `mortality_nonsmoker/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(capital, value, color = values, fill = values), alpha = 0.4)
```

```{r}
# BMI with sex
t1 <- insurance_train %>%
  mutate(BMI = round(BMI)) %>%
  group_by(BMI) %>%
  mutate(
    mortality_man = sum(mortality == "yes" & sex == "man"),
    mortality_woman = sum(mortality == "yes" & sex == "woman"),
    ) %>%
  select(BMI, mortality_man, mortality_woman) %>%
  distinct() %>%
  arrange(BMI)

t1 <- t1[-8, ]

t_man <- insurance_train %>%
  mutate(BMI = round(BMI)) %>%
  group_by(BMI) %>%
  filter(sex == "man") %>%
  mutate(exp_man = sum(exp)) %>%
  select(BMI, exp_man) %>%
  distinct() %>%
  arrange(BMI)

t_woman <- insurance_train %>%
  mutate(BMI = round(BMI)) %>%
  group_by(BMI) %>%
  filter(sex == "woman") %>%
  mutate(exp_woman = sum(exp)) %>%
  select(BMI, exp_woman) %>%
  distinct() %>%
  arrange(BMI)

t_woman <- t_woman[-8, ]

t <- t1 %>%
  bind_cols(
    exp_man = t_man$exp_man,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_woman = t_woman$exp_woman,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_man/exp` = mortality_man/exp_man,
    `mortality_woman/exp` = mortality_woman/exp_woman
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_man/exp`, `mortality_woman/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(BMI, value, color = values, fill = values), alpha = 0.4)

t %>%
  ggplot(aes(BMI, value, color = values, fill = values)) +
  geom_point() +
  scale_fill_manual(values = group_colors) +
  geom_smooth(method = "loess") +
  scale_y_continuous(labels = percent_format())
```

```{r}
# BMI with smoker
t1 <- insurance_train %>%
  mutate(BMI = round(BMI, 1)) %>%
  group_by(BMI) %>%
  mutate(
    mortality_smoker = sum(mortality == "yes" & smoker == "yes"),
    mortality_nonsmoker = sum(mortality == "yes" & smoker == "no"),
    ) %>%
  select(BMI, mortality_smoker, mortality_nonsmoker) %>%
  distinct() %>%
  arrange(BMI)

t1 <- t1[-8, ]

t_yes <- insurance_train %>%
  mutate(BMI = round(BMI, 1)) %>%
  group_by(BMI) %>%
  filter(smoker == "yes") %>%
  mutate(exp_yes = sum(exp)) %>%
  select(BMI, exp_yes) %>%
  distinct() %>%
  arrange(BMI)

t_no <- insurance_train %>%
  mutate(BMI = round(BMI, 1)) %>%
  group_by(BMI) %>%
  filter(smoker == "no") %>%
  mutate(exp_no = sum(exp)) %>%
  select(BMI, exp_no) %>%
  distinct() %>%
  arrange(BMI)

t_no <- t_no[-8, ]

t <- t1 %>%
  bind_cols(
    exp_yes = t_yes$exp_yes,
    .name_repair = "unique"
    ) %>%
  bind_cols(
    exp_no = t_no$exp_no,
    .name_repair = "unique"
  ) %>%
  mutate(
    `mortality_smoker/exp` = mortality_smoker/exp_yes,
    `mortality_nonsmoker/exp` = mortality_nonsmoker/exp_no
    )

t <- t %>%
  pivot_longer(cols = c(`mortality_smoker/exp`, `mortality_nonsmoker/exp`), names_to = "values")

t %>%
  ggplot() +
  geom_point(aes(BMI, value, color = values, fill = values), alpha = 0.4)
```

####

```{r}
# GAM
insurance_train_prep <- 
  recipe(mortality ~ sex + smoker +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  themis::step_upsample(mortality, over_ratio = 0.01, seed = 123) %>%
  step_log(capital, BMI, base = 10) %>%
  prep() %>% juice()

gam_model <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps") +
                   s(duration, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = smoker) +
                   s(capital, bs = "ps", by = sex) +
                   s(BMI, bs = "ps", by = sex),
                 data = insurance_train_prep, family = "binomial", method = "REML", select = F)
summary(gam_model)

t <- Sys.time()
gam_model <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = smoker) +
                   s(capital, bs = "ps", by = sex) +
                   s(BMI, bs = "ps", by = sex),
                 data = insurance_train_prep, family = "binomial", method = "REML", select = T)
# saveRDS(gam_model, "../01_mortality_model/results_mortality/gam_model_mortality.RDS")
t1 <- Sys.time()
t1 - t
gam_model <- gam_model_mortality
itsadug::gamtabs(gam_model)
summary(gam_model)
plot(gam_model, ylim = c(-5,5), rug = T)
# Parametric coefficients:
#             Estimate Std. Error  z value Pr(>|z|)    
# (Intercept) -4.94215    0.04004 -123.420   <2e-16 ***
# sexwoman    -0.91165    0.08209  -11.106   <2e-16 ***
# smokeryes    0.04446    0.07601    0.585    0.559    
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Approximate significance of smooth terms:
#                              edf Ref.df   Chi.sq  p-value    
# s(actuarial_age):sexman   5.7830      9 1575.682  < 2e-16 ***
# s(actuarial_age):sexwoman 7.2456      9  461.767  < 2e-16 ***
# s(duration):sexman        7.4831      9   84.584  < 2e-16 ***
# s(duration):sexwoman      4.7061      9   27.212 1.83e-07 ***
# s(duration):smokerno      0.9659      9   10.974 7.11e-07 ***
# s(duration):smokeryes     7.8842      9  164.985  < 2e-16 ***
# s(capital):sexman         6.9941      9  145.811  < 2e-16 ***
# s(capital):sexwoman       0.6583      8    1.872   0.0910 .  
# s(BMI):sexman             2.1023      9    8.084   0.0125 *  
# s(BMI):sexwoman           4.0590      8  106.777  < 2e-16 ***
```

```{r}
pred <- predict(gam_model, insurance_train_prep, type = "response")
predictions <- tibble(
  truth = insurance_train_prep$mortality,
  prob = pred
)
autoplot(pr_curve(predictions, truth, prob, event_level = "second"))
ggsave("roc1.jpg")
```

```{r}
plot_gam <- getViz(gam_model)
```

```{r}
plot(sm(plot_gam, 1)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 2)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 3)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 4)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 5)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 6)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  theme_bw()
plot(sm(plot_gam, 7)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  scale_x_continuous(labels = function(x) scales::scientific_format()(10^x)) +
  theme_bw()
plot(sm(plot_gam, 8)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  scale_x_continuous(labels = function(x) scales::scientific_format()(10^x)) +
  theme_bw()
plot(sm(plot_gam, 9)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  scale_x_continuous(labels = function(x) round(10^x, 1)) +
  theme_bw()
plot(sm(plot_gam, 10)) +
  l_fitLine(colour = "indianred", size = 1, alpha = 0.8) +
  l_rug(mapping = aes(x=x), alpha = 0.4) +
  l_ciLine(mul = 5, colour = "indianred", linetype = 2, size = 1) +
  scale_x_continuous(labels = function(x) round(10^x, 1)) +
  theme_bw()

a <- c(3, 4, 5, 6)
scales::scientific_format()(10^a)
parse(text=gsub("e", " %*% 10^", a))
```

```{r}
insurance_train <- merge(insurance_train, prima, by = "actuarial_age") %>% 
  mutate(qx = if_else(sex == "man", qx_men, qy_women)) %>% 
  select(-qx_men, -qy_women) %>% 
  mutate(prima = qx * capital) %>% 
  mutate(prima = 1.15*prima)
```


```{r}
# Best threshold for kappa
insurance_train_prep <- 
  recipe(mortality ~ sex + smoker + 
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  step_log(capital, BMI, base = 10) %>%
  prep() %>% juice()

prob <- predict(gam_model, newdata = insurance_train_prep, 
                type = "response")
prob <- as_tibble(prob) %>%
  mutate(mortality = insurance_train_prep$mortality)

roc_c <- prob %>%
  roc_curve(mortality, value, event_level = "second")

thresholds <- roc_c %>% select(.threshold) %>% pull()
length(thresholds)
length(which(thresholds <= 0.08966623))
thresholds <- quantile(thresholds, probs = seq(0.98, 0.999, 0.0001))

thresholds <- roc_c %>% select(.threshold) %>% pull()
thresholds <- thresholds[(255923-300):(255923+100)]
kappa <- seq(1, length(thresholds))

for (i in 1:length(thresholds)){
  confusion_matrix <- prob %>%
    mutate(pred_class = as.factor(ifelse(value > thresholds[i],
                                         "yes", "no"))) %>%
    conf_mat(truth = mortality,
             estimate = pred_class)
  
  kappa[i] <- confusion_matrix %>% 
    summary(event_level = "second") %>%
    filter(.metric == "kap") %>% 
    dplyr::select(.estimate) %>% 
    pull()
}
which.max(kappa)
thres <- 0.08966623
confusion_matrix <- prob %>%
    mutate(pred_class = as.factor(ifelse(value > thres,
                                         "yes", "no"))) %>%
    conf_mat(truth = mortality,
             estimate = pred_class)
confusion_matrix
confusion_matrix %>% 
    summary(event_level = "second") %>%
    filter(.metric == "kap") %>% 
    dplyr::select(.estimate) %>% 
    pull()

plot(kappa)
```

```{r}
# Best threshold for Fscore
insurance_train_prep <- 
  recipe(mortality ~ sex + smoker + prima +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  step_log(capital, BMI, base = 10) %>%
  prep() %>% juice()

prob <- predict(gam_model, newdata = insurance_train_prep, 
                type = "response")

prob <- as_tibble(prob) %>%
  mutate(mortality = insurance_train_prep$mortality,
         capital = insurance_train$capital,
         prima = insurance_train$prima)

prec_rec_curve <- prob %>%
  pr_curve(mortality, value, event_level = "second")

b0 <- 0
b_beta <- 1
for (beta in 1:20){
  thres <- prec_rec_curve %>%
  mutate(sum = (1 + beta^2)*(recall * precision)/((beta^2*precision) + recall)) %>%
  filter(sum == max(sum, na.rm = T)) %>%
  select(.threshold) %>%
  pull()

  results <- prob %>%
  mutate(pred_class = as.factor(ifelse(value > thres,
                                       "yes", "no"))) %>% 
  mutate(
    capital = as.double(capital),
    result = case_when(
      mortality == "no" & pred_class == "no" ~ prima,
      mortality == "no" & pred_class == "yes" ~ -prima,
      mortality == "yes" & pred_class == "no" ~ -capital+prima,
      mortality == "yes" & pred_class == "yes" ~ capital-prima,
  )) %>% 
  select(result) %>% sum
  
  if (results > b0){
    b0 <- results
    b_beta <- beta
  }
  
}
b0
b_beta


thres <- prec_rec_curve %>%
  mutate(sum = (1 + b_beta^2)*(recall * precision)/((b_beta^2*precision) + recall)) %>%
  filter(sum == max(sum, na.rm = T)) %>%
  select(.threshold) %>%
  pull()

prob <- prob %>%
  mutate(pred_class = as.factor(ifelse(value > thres,
                                     "yes", "no"))) 

prob %>% 
  conf_mat(mortality, pred_class) %>% 
  summary(beta = 2, event_level = "second")

# Facturación real
sum(prob$prima)
# Siniestralidad real
sum(prob$capital[which(prob$mortality == "yes")])

# Facturación teórica
prob %>% 
  summarise(n_prima = value * capital) %>% 
  sum

# Siniestralidad teórica

```

```{r}
# Best threshold for F-score
insurance_train_prep <- 
  recipe(mortality ~ sex + smoker +
                   actuarial_age + duration + capital + BMI,
         data = insurance_train) %>%
  step_log(capital, BMI, base = 10) %>%
  prep() %>% juice()
insurance_test_prep <- 
  recipe(mortality ~ sex + smoker +
                   actuarial_age + duration + capital + BMI,
         data = insurance_test) %>%
  step_log(capital, BMI, base = 10) %>%
  prep() %>% juice()

prob <- predict(gam_model, newdata = insurance_train_prep, 
                type = "response")

prob <- as_tibble(prob) %>%
  mutate(mortality = insurance_train_prep$mortality)

prec_rec_curve <- prob %>%
  pr_curve(mortality, value, event_level = "second")

b_beta <- 8
thres <- prec_rec_curve %>%
  mutate(sum = (1 + b_beta^2)*(recall * precision)/((b_beta^2*precision) + recall)) %>%
  filter(sum == max(sum, na.rm = T)) %>%
  select(.threshold) %>%
  pull()

# thres <- 0.04014321
confusion_matrix <- prob %>%
  mutate(pred_class = as.factor(ifelse(value > thres,
                                       "yes", "no"))) %>%
  conf_mat(truth = mortality,
           estimate = pred_class)
confusion_matrix

# TEST SET
prob <- predict(gam_model, newdata = insurance_test_prep, 
                type = "response")

prob <- as_tibble(prob) %>%
  mutate(mortality = insurance_test_prep$mortality)

prob %>%
  mutate(pred_class = as.factor(ifelse(value > 0.04014321,
                                       "yes", "no"))) %>%
  conf_mat(truth = mortality,
           estimate = pred_class) %>%
  summary(event_level = "second", beta = 8)

confusion_matrix %>%
  summary(event_level = "second")
```



```{r}
gam_model <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps") + s(duration, bs = "ps") +
                   s(capital, bs = "ps") + s(BMI, bs = "ps") +
                   s(duration, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = smoker) +
                   s(capital, bs = "ps", by = sex) +
                   s(BMI, bs = "ps", by = sex),
                 data = insurance_train_prep, family = "binomial", method = "REML", select = T)
summary(gam_model)
# saveRDS(gam_model, "../01_mortality_model/results_mortality/gam_model.RDS")

# Parametric coefficients:
#             Estimate Std. Error  z value Pr(>|z|)    
# (Intercept) -4.91023    0.03809 -128.928   <2e-16 ***
# sexwoman    -0.95785    0.06505  -14.724   <2e-16 ***
# smokeryes    0.04702    0.07596    0.619    0.536    
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Approximate significance of smooth terms:
#                            edf Ref.df   Chi.sq  p-value    
# s(actuarial_age)      3.720374      9 1986.664  < 2e-16 ***
# s(duration)           0.009843      9    0.009 0.000194 ***
# s(capital)            0.043252      9    0.041 0.000461 ***
# s(BMI)                0.003195      9    0.003 0.010083 *  
# s(duration):sexman    7.482038      9   85.939  < 2e-16 ***
# s(duration):sexwoman  4.741846      9   29.893  < 2e-16 ***
# s(duration):smokerno  0.951182      9   11.469  8.1e-07 ***
# s(duration):smokeryes 7.884977      9  166.362  < 2e-16 ***
# s(capital):sexman     6.972616      9  117.955  < 2e-16 ***
# s(capital):sexwoman   3.306112      8    9.748 0.001672 ** 
# s(BMI):sexman         2.128135      9    8.654 0.007842 ** 
# s(BMI):sexwoman       4.072868      8  111.462  < 2e-16 ***
plot(gam_model, ylim = c(-5, 5))
```

```{r}
prob <- predict(gam_model, newdata = insurance_train_prep, type = "response")

prob <- as_tibble(prob) %>%
  mutate(mortality = insurance_train_prep$mortality)

prec_rec_curve <- prob %>%
  pr_curve(mortality, value, event_level = "second")

thres <- prec_rec_curve %>%
  mutate(sum = (5)*(recall * precision)/((4*precision) + recall)) %>%
  filter(sum == max(sum, na.rm = T)) %>%
  select(.threshold) %>%
  pull()
# thres <- 0.03884555

confusion_matrix <- prob %>%
  mutate(pred_class = as.factor(ifelse(value > thres,
                                       "yes", "no"))) %>%
  conf_mat(truth = mortality,
           estimate = pred_class)

confusion_matrix %>%
  summary(event_level = "second", beta = 2)
```




```{r}
set.seed(23)
# Recipe
rec <- 
  recipe(mortality ~ sex + smoker + actuarial_age + BMI_factor_1 + capital, 
         data = insurance_train) %>%
  themis::step_upsample(mortality, over_ratio = 0.01, seed = 123) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_bs(capital, deg_free = 3)

rec %>% prep() %>% juice

# Model
glm_model <- logistic_reg() %>%
  set_engine("glm")

# Workflow
glm_wflow <- workflow() %>%
  add_model(glm_model) %>%
  add_recipe(rec)

glm_fit <- fit(glm_wflow, insurance_train)
tidy(glm_fit)
```


# Interaction terms


# Preprocessing steps:

After seeing some description of the variables, their effect on the model and their interactions, the preprocessing steps we are going to follow are:

1. A method of resampling for the unbalanced class (e.g. `step_upsample(mortality)`).

2. Convert to dummy variables all categorical since it is usually useful for the majority of models.

3. We will add the interactions between the final variables.

Also, for some models, it is mandatory to scale and center the variables, but for the moment, we are not going to do it.
