==> 01_mortality_model/4_ModelTuning_mortality.Rmd <==
---
title: '**Mortality: Model tuning and selection**'
author: '*Roberto Jesús Alcaraz Molina*'
date: "14/04/2021"
output:
  html_document:
    toc: T
    fig_caption: yes
header-includes: 
- \usepackage{float}
- \usepackage{amsbsy}
- \usepackage{amsmath}
- \usepackage{graphicx}
- \usepackage{subfig}
- \usepackage{booktabs}
---

```{r include=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      echo = F, 
                      message = FALSE,
                      fig.pos="H", 
                      fig.align="center",
                      fig.width=15,
                      cache=FALSE, error = TRUE)
```

```{r, echo = F}
# install.packages("pacman")
pacman::p_load(tidymodels, workflowsets, tune, patchwork, FactoMineR, doParallel, usemodels, mgcv)
theme_set(theme_bw())

# Models packages
pacman::p_load(ranger)
library(tidyverse, attach.required = T)
```

## Code

```{r, eval = T}
insurance <- readRDS("../00_data/insurance.RDS")

insurance <- insurance %>%
  filter(actuarial_age >= 21, actuarial_age <= 61) %>%
  filter(duration <= 13)

insurance$capital <- as_double(insurance$capital)


set.seed(123)

insurance_split <- initial_split(insurance, prop = 0.7, strata = mortality)
insurance_train <- training(insurance_split)
insurance_test  <- testing(insurance_split)

set.seed(23)
insurance_folds <- vfold_cv(insurance_train, v = 10, 
                            strata = mortality)
```

We are going to compare the following models: 

```{r}
grid_ctrl <- control_grid(
  save_pred = TRUE, 
  save_workflow = TRUE, 
  event_level = "second"
  )

f_meas_2 <- function(data, truth, estimate, na_rm = TRUE, ...){
  f_meas(
    data = data,
    truth = !! rlang::enquo(truth),
    estimate = !! rlang::enquo(estimate),
    # set beta = 2
    beta = 8,
    na_rm = TRUE,
    ...
  )
}

f_meas_2 <- new_class_metric(f_meas_2, direction = "maximize")

# It works
# prob %>%
#     mutate(pred_class = as.factor(ifelse(value > t[i], "yes", "no"))) %>%
#     f_meas(mortality, pred_class, beta = 2, event_level = "second")
# prob %>%
#     mutate(pred_class = as.factor(ifelse(value > t[i], "yes", "no"))) %>%
#     f_meas_2(mortality, pred_class)

my_metrics = metric_set(f_meas_2, kap, sensitivity, specificity)
```


# Models

```{r}
kknn_rec <- 
  recipe(formula = mortality ~ sex + smoker + duration + actuarial_age + BMI + 
    capital, data = insurance_train) %>% 
  themis::step_downsample(mortality, under_ratio = tune(), seed = 123) %>% 
  step_log(capital, BMI, base = 10) %>%
  step_dummy(all_nominal(), -all_outcomes()) %>% 
  step_interact(terms = ~ duration:sex_woman + duration:smoker_yes + 
                  actuarial_age:sex_woman + actuarial_age:smoker_yes + 
                  BMI:sex_woman + BMI:smoker_yes +                  
                  capital:sex_woman + capital:smoker_yes) %>%
  step_zv(all_predictors()) %>% 
  step_normalize(all_predictors(), -all_nominal()) 

kknn_model <- 
  nearest_neighbor(neighbors = tune(), dist_power = tune()) %>% 
  set_mode("classification") %>% 
  set_engine("kknn") 

kknn_wf <- 
  workflow() %>% 
  add_recipe(kknn_rec) %>% 
  add_model(kknn_model) 
```

```{r}
svm_rec <- 
  recipe(mortality ~ sex + smoker + duration + actuarial_age + BMI + capital, 
         data = insurance_train) %>%
  themis::step_downsample(mortality, under_ratio = tune(), seed = 123) %>% 
  step_log(capital, BMI, base = 10) %>%
  step_normalize(all_numeric_predictors()) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_interact(terms = ~ duration:sex_woman + duration:smoker_yes +  
                  actuarial_age:sex_woman + actuarial_age:smoker_yes +
                  BMI:sex_woman + BMI:smoker_yes +               
                  capital:sex_woman + capital:smoker_yes) %>%
  step_zv(all_predictors()) %>%
  step_corr(all_predictors())

svm_model <-
  svm_poly(cost = tune(), degree = tune(), scale_factor = tune()) %>%
  set_engine('kernlab') %>%
  set_mode('classification')

svm_wf <- 
  workflow() %>%
  add_recipe(svm_rec) %>%
  add_model(svm_model)
```


```{r}
# Decision Tree
dt_rec <- 
  recipe(mortality ~ sex + smoker + duration + actuarial_age + BMI + capital, 
         data = insurance_train) %>%
  themis::step_downsample(mortality, under_ratio = tune(), seed = 123) %>% 
  step_log(capital, BMI, base = 10) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_interact(terms = ~ duration:sex_woman + duration:smoker_yes +
                  actuarial_age:sex_woman + actuarial_age:smoker_yes +
                  BMI:sex_woman + BMI:smoker_yes +
                  capital:sex_woman + capital:smoker_yes)

dt_model <- 
  decision_tree(cost_complexity = tune(), tree_depth = tune(), min_n = tune()) %>% 
  set_mode("classification") %>% 
  set_engine("rpart")

dt_wf <- 
  workflow() %>%
  add_recipe(dt_rec) %>%
  add_model(dt_model)
```

```{r}
# Random Forest
rf_rec <- 
  recipe(mortality ~ sex + smoker + duration + actuarial_age + BMI + capital, 
         data = insurance_train) %>%
  themis::step_downsample(mortality, under_ratio = tune(), seed = 123) %>% 
  step_log(capital, BMI, base = 10) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_interact(terms = ~ duration:sex_woman + duration:smoker_yes +
                  actuarial_age:sex_woman + actuarial_age:smoker_yes +
                  BMI:sex_woman + BMI:smoker_yes +
                  capital:sex_woman + capital:smoker_yes)

rf_model <- 
  rand_forest(mtry = tune(), min_n = tune(), trees = 1000) %>% 
  set_mode("classification") %>% 
  set_engine("ranger")

rf_wf <- 
  workflow() %>%
  add_recipe(rf_rec) %>%
  add_model(rf_model)
```


```{r}
mlp_rec <- 
  recipe(mortality ~ sex + smoker + duration + actuarial_age + BMI + capital, 
         data = insurance_train) %>%
  themis::step_downsample(mortality, under_ratio = tune(), seed = 123) %>% 
  step_log(capital, BMI, base = 10) %>%
  step_normalize(all_numeric_predictors()) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_interact(terms = ~ duration:sex_woman + duration:smoker_yes +
                  actuarial_age:sex_woman + actuarial_age:smoker_yes + 
                  BMI:sex_woman + BMI:smoker_yes +                
                  capital:sex_woman + capital:smoker_yes) %>%
  step_zv(all_predictors()) %>%
  step_corr(all_predictors())

mlp_model <-
  mlp(hidden_units = tune(), epochs = tune(), penalty = tune()) %>%
  set_engine('nnet') %>%
  set_mode('classification')

mlp_wf <- 
  workflow() %>%
  add_recipe(mlp_rec) %>%
  add_model(mlp_model)
```


# Parameter tuning 

```{r, eval=F}
# KKNN
set.seed(123)
kknn_param <- kknn_wf %>% 
  parameters() %>% 
  update(
    neighbors = neighbors(c(1, 21)),
    under_ratio = under_ratio(c(4, 50))
  )
  
max_ent_kknn <- grid_max_entropy(kknn_param, size = 30)
max_ent_kknn$dist_power <- ifelse(max_ent_kknn$dist_power >= 1.5, 2, 1)

t <- Sys.time()
cl <- makePSOCKcluster(8)
registerDoParallel(cl)

kknn_wf_tune <- kknn_wf %>%
  tune_grid(
    resamples = insurance_folds,
    grid = max_ent_kknn,
    control = grid_ctrl,
    metrics = my_metrics
    )

stopCluster(cl)
saveRDS(kknn_wf_tune, "../01_mortality_model/results_mortality/kknn_wf_tune.RDS")
t1 <- Sys.time()
t1 - t

# SVM
set.seed(123)
svm_param <- svm_wf %>% 
  parameters() %>% 
  update(
    under_ratio = under_ratio(c(4, 50))
  )
  
max_ent_svm <- grid_max_entropy(svm_param, size = 30)

t <- Sys.time()
cl <- makePSOCKcluster(8)
registerDoParallel(cl)

svm_wf_tune <- svm_wf %>%
  tune_grid(
    resamples = insurance_folds,
    grid = max_ent_svm,
    control = grid_ctrl,
    metrics = my_metrics
    )

stopCluster(cl)
saveRDS(svm_wf_tune, "../01_mortality_model/results_mortality/svm_wf_tune.RDS")
t2 <- Sys.time()
t2 - t

# Decision tree
set.seed(123)

dt_param <- dt_wf %>%
  parameters() %>%
  update(
    under_ratio = under_ratio(c(4, 50))
    )

max_ent_dt <- grid_max_entropy(dt_param, size = 30)

t <- Sys.time()
cl <- makePSOCKcluster(8)
registerDoParallel(cl)

dt_wf_tune <- dt_wf %>%
  tune_grid(
    resamples = insurance_folds,
    grid = max_ent_dt,
    control = grid_ctrl,
    metrics = my_metrics
    )

saveRDS(dt_wf_tune, "../01_mortality_model/results_mortality/dt_wf_tune.RDS")
stopCluster(cl)

t3 <- Sys.time()
t3 - t

# Random Forest
set.seed(123)

rf_param <- rf_wf %>%
  parameters() %>%
  update(
    mtry = finalize(mtry(c(1, 15))),
    under_ratio = under_ratio(c(4, 50))
    )

max_ent_rf <- grid_max_entropy(rf_param, size = 30)

t <- Sys.time()
cl <- makePSOCKcluster(8)
registerDoParallel(cl)

rf_wf_tune <- rf_wf %>%
  tune_grid(
    resamples = insurance_folds,
    grid = max_ent_rf,
    control = grid_ctrl,
    metrics = my_metrics
    )

saveRDS(rf_wf_tune, "../01_mortality_model/results_mortality/rf_wf_tune.RDS")
stopCluster(cl)

t4 <- Sys.time()
t4 - t

# MLP
set.seed(123)
mlp_param <- mlp_wf %>% 
  parameters() %>% 
  update(
    under_ratio = under_ratio(c(4, 50))
  )

max_ent_mlp <- grid_max_entropy(mlp_param, size = 30)

t <- Sys.time()
cl <- makePSOCKcluster(8)
registerDoParallel(cl)

mlp_wf_tune <- mlp_wf %>%
  tune_grid(
    resamples = insurance_folds,
    grid = max_ent_mlp,
    control = grid_ctrl,
    metrics = my_metrics
    )

stopCluster(cl)
saveRDS(mlp_wf_tune, "../01_mortality_model/results_mortality/mlp_wf_tune.RDS")
t5 <- Sys.time()
t5 - t
```

```{r}
kknn_wf_tune <- readRDS("results_mortality/kknn_wf_tune.RDS") 
svm_wf_tune <- readRDS("results_mortality/svm_wf_tune.RDS") 
dt_wf_tune <- readRDS("results_mortality/dt_wf_tune.RDS") 
rf_wf_tune <- readRDS("results_mortality/rf_wf_tune.RDS") 
mlp_wf_tune <- readRDS("results_mortality/mlp_wf_tune.RDS") 
```

```{r}
wflow_set <- as_workflow_set(
  kknn = kknn_wf_tune, 
  svm = svm_wf_tune,
  dt = dt_wf_tune,
  rf = rf_wf_tune,
  mlp = mlp_wf_tune
  )


wflow_set %>% collect_metrics() %>% 
  filter(.metric == "f_meas") %>% 
  arrange(desc(mean)) # %>% 
  # filter(wflow_id == "mlp")

wflow_set %>% collect_metrics() %>% 
  filter(wflow_id == "kknn", .config == "Preprocessor11_Model1")

mlp_wf_tune %>%
  collect_metrics() %>%
  filter(.metric == "f_meas") %>%
  arrange(desc(mean)) 

autoplot(wflow_set)
autoplot(wflow_set, rank_metric = "kap", select_best = T)

ggsave("../01_mortality_model/results_mortality/mort.png", width = 10)
```

```{r}
# GAM
n <- nrow(insurance_folds)

f8 <- rep(0, 1, n)
kappa <- rep(0, 1, n)
sensit <- rep(0, 1, n)
specif <- rep(0, 1, n)

t <- Sys.time()

for (i in 1:n){
  
  analysis_data <- analysis(insurance_folds$splits[[i]])
  analysis_data <- 
    recipe(mortality ~ sex + smoker + actuarial_age + duration + capital + BMI,
           data = analysis_data) %>%
    themis::step_downsample(mortality, under_ratio = 20, seed = 123) %>%
    step_log(capital, BMI, base = 10) %>% 
    prep() %>% juice()
  
  validation_data <- assessment(insurance_folds$splits[[i]])
  
  gam_model <- gam(mortality ~ sex + smoker +
                   s(actuarial_age, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = sex) +
                   s(duration, bs = "ps", by = smoker) +
                   s(capital, bs = "ps", by = sex) +
                   s(BMI, bs = "ps", by = sex),
                 data = analysis_data, family = "binomial", method = "REML", select = T)
  
  prob <- predict(gam_model, newdata = validation_data, type = "response")

  prob <- as_tibble(prob) %>%
    mutate(mortality = validation_data$mortality)
  
  confusion_matrix <- prob %>%
    mutate(pred_class = as.factor(ifelse(value > 0.01173517, "yes", "no"))) %>%
    mutate(pred_class = factor(pred_class, levels = c("no", "yes"), labels = c("no", "yes"))) %>% 
    conf_mat(truth = mortality, estimate = pred_class)
  
  f8[i] <- confusion_matrix %>% summary(event_level = "second", beta = 8) %>%
    filter(.metric == "f_meas") %>% dplyr::select(.estimate) %>% pull()
  kappa[i] <- confusion_matrix %>% summary(event_level = "second") %>%
    filter(.metric == "kap") %>% dplyr::select(.estimate) %>% pull()
  sensit[i] <- confusion_matrix %>% summary(event_level = "second") %>%
    filter(.metric == "sens") %>% dplyr::select(.estimate) %>% pull()
  specif[i] <- confusion_matrix %>% summary(event_level = "second") %>%
    filter(.metric == "spec") %>% dplyr::select(.estimate) %>% pull()
}
t1 <- Sys.time()
t1 - t

mean(f8)
mean(kappa)
mean(sensit)
mean(specif)
```


```{r}
# MLP
mlp_param <- mlp_wf_tune %>%
  collect_metrics() %>%
  filter(.metric == "f_meas") %>%
  arrange(desc(mean)) %>%
  slice(1) %>%
  dplyr::select(hidden_units, penalty, epochs, under_ratio)

mlp_wf <- mlp_wf %>%
  finalize_workflow(mlp_param)

mlp_final_wf <- mlp_wf %>%
  last_fit(insurance_split)

mlp_final_wf %>%
  collect_predictions() %>%
  conf_mat(truth = mortality,
           estimate = .pred_class) %>%
  summary(event_level = "second", beta = 2)

# DT
dt_param <- dt_wf_tune %>%
  collect_metrics() %>%
  filter(.metric == "f_meas") %>%
  arrange(desc(mean)) %>%
  slice(1) %>%
  dplyr::select(cost_complexity, tree_depth, min_n, under_ratio)

dt_wf <- dt_wf %>%
  finalize_workflow(dt_param)

dt_final_wf <- dt_wf %>%
  fit(insurance_train)
rpart.plot::rpart.plot(dt_final_wf$fit$fit$fit, roundint=FALSE, 
                       fallen.leaves = T, cex = 0.65)
ggsave("dt.jpg")

pred <- predict(dt_final_wf, insurance_train_prep)

pred %>%
  bind_cols(mortality = insurance_test_prep$mortality) %>% 
  conf_mat(truth = mortality,
           estimate = .pred_class) %>%
  summary(event_level = "second", beta = 8)

# KKNN
kknn_param <- kknn_wf_tune %>%
  collect_metrics() %>%
  filter(.metric == "f_meas") %>%
  arrange(desc(mean)) %>%
  slice(1) %>%
  dplyr::select(neighbors, dist_power, under_ratio)

kknn_wf <- kknn_wf %>%
  finalize_workflow(kknn_param)

kknn_final_wf <- kknn_wf %>%
  last_fit(insurance_split)

kknn_final_wf %>%
  collect_predictions() %>%
  conf_mat(truth = mortality,
           estimate = .pred_class) %>%
  summary(event_level = "second", beta = 2)
cost
```
